Nick's Roguelike Game
Followed along and made it to 1:24:42 in the video. Converted it from my online compiler to visual studio when I got home from work but it still requires you to move and click enter after every move. Also there's a bug where it says invalid input now when moving the character.(probably due to pressing the enter key) Game also crashes when you go and attack a skeleton.